+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+* BREAKING CHANGE: Represent `sbyte` using number instead of string.
+* BREAKING CHANGE: Represent `byte` using number instead of string.
+* BREAKING CHANGE: Represent `int16` using number instead of string.
+* BREAKING CHANGE: Represent `uint16` using number instead of string.
+
+
 ## 6.0.0 - 2021-07-06
 
 ### Added